Safari-611.1.10/LayoutTests/fast/dom/append-child-with-mutation-event-removal-and-circular-shadow-insertion-expected.txt

This tests re-parenting a child of the document fragment during an insertion so as to create a circular node tree.
WebKit should detect this case and throw HierarchyRequestError.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS container.appendChild(fragment) threw exception HierarchyRequestError: The operation would yield an incorrect node tree..
PASS container.parentNode is childShadowRoot
PASS child.parentNode is null
PASS successfullyParsed is true

TEST COMPLETE
